@@ -33,14 +33,19 @@ namespace Examine
         /// <inheritdoc/>
 		public long TotalItemCount => 0;
 
+
+#pragma warning disable IDE0060 // Remove unused parameter
         /// <inheritdoc/>
         public IEnumerable<ISearchResult> Skip(int skip)
-		{
+#pragma warning restore IDE0060 // Remove unused parameter
+        {
 			return Enumerable.Empty<ISearchResult>();
 		}
 
         /// <inheritdoc/>
+#pragma warning disable IDE0060 // Remove unused parameter
         public IEnumerable<ISearchResult> SkipTake(int skip, int? take = null)
+#pragma warning restore IDE0060 // Remove unused parameter
         {
 			return Enumerable.Empty<ISearchResult>();
 		}
