@@ -9,35 +9,27 @@ namespace Examine.Search
     public interface IFacetOperations : IQueryExecutor
     {
         /// <summary>
-        /// Add a facet string to the current query
-        /// </summary>
-        /// <param name="field"></param>
-        /// <param name="facetConfiguration"></param>
-        /// <returns></returns>
-        IFacetOperations Facet(string field, Action<IFacetQueryField>? facetConfiguration = null);
-
-        /// <summary>
         /// Add a facet string to the current query, filtered by a single value or multiple values
         /// </summary>
         /// <param name="field"></param>
         /// <param name="facetConfiguration"></param>
         /// <param name="values"></param>
         /// <returns></returns>
-        IFacetOperations Facet(string field, Action<IFacetQueryField>? facetConfiguration = null, params string[] values);
+        IFacetOperations FacetString(string field, Action<IFacetQueryField>? facetConfiguration = null, params string[] values);
 
         /// <summary>
         /// Add a range facet to the current query
         /// </summary>
-        IFacetOperations Facet(string field, params DoubleRange[] doubleRanges);
+        IFacetOperations FacetDoubleRange(string field, params DoubleRange[] doubleRanges);
 
         /// <summary>
         /// Add a range facet to the current query
         /// </summary>
-        IFacetOperations Facet(string field, params FloatRange[] floatRanges);
+        IFacetOperations FacetFloatRange(string field, params FloatRange[] floatRanges);
 
         /// <summary>
         /// Add a range facet to the current query
         /// </summary>
-        IFacetOperations Facet(string field, params Int64Range[] longRanges);
+        IFacetOperations FacetLongRange(string field, params Int64Range[] longRanges);
     }
 }
