@@ -9,7 +9,6 @@ using Lucene.Net.Analysis;
 using Lucene.Net.Facet;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.DependencyInjection.Extensions;
-using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Options;
 
 namespace Examine
@@ -19,48 +18,83 @@ namespace Examine
     /// </summary>
     public static class ServicesCollectionExtensions
     {
-
         /// <summary>
         /// Registers a file system based Lucene Examine index
         /// </summary>
+        [Obsolete("To remove in Examine V5")]
         public static IServiceCollection AddExamineLuceneIndex(
             this IServiceCollection serviceCollection,
             string name,
-            FieldDefinitionCollection? fieldDefinitions = null,
-            Analyzer? analyzer = null,
-            IValueSetValidator? validator = null,
-            IReadOnlyDictionary<string, IFieldValueTypeFactory>? indexValueTypesFactory = null,
-            FacetsConfig? facetsConfig = null)
-            => serviceCollection.AddExamineLuceneIndex<LuceneIndex>(name, fieldDefinitions, analyzer, validator, indexValueTypesFactory, facetsConfig);
+            FieldDefinitionCollection fieldDefinitions = null,
+            Analyzer analyzer = null,
+            IValueSetValidator validator = null,
+            IReadOnlyDictionary<string, IFieldValueTypeFactory> indexValueTypesFactory = null)
+            => serviceCollection.AddExamineLuceneIndex<LuceneIndex>(name, fieldDefinitions, analyzer, validator, indexValueTypesFactory);
 
         /// <summary>
         /// Registers a file system based Lucene Examine index
         /// </summary>
+        [Obsolete("To remove in Examine V5")]
         public static IServiceCollection AddExamineLuceneIndex<TIndex>(
             this IServiceCollection serviceCollection,
             string name,
-            FieldDefinitionCollection? fieldDefinitions = null,
-            Analyzer? analyzer = null,
-            IValueSetValidator? validator = null,
-            IReadOnlyDictionary<string, IFieldValueTypeFactory>? indexValueTypesFactory = null,
-            FacetsConfig? facetsConfig = null)
+            FieldDefinitionCollection fieldDefinitions = null,
+            Analyzer analyzer = null,
+            IValueSetValidator validator = null,
+            IReadOnlyDictionary<string, IFieldValueTypeFactory> indexValueTypesFactory = null)
             where TIndex : LuceneIndex
-            => serviceCollection.AddExamineLuceneIndex<TIndex, FileSystemDirectoryFactory>(name, fieldDefinitions, analyzer, validator, indexValueTypesFactory, facetsConfig);
+        {
+            Action<ExamineLuceneIndexConfiguration<TIndex, FileSystemDirectoryFactory>> config = opt =>
+            {
+                opt.FieldDefinitions = fieldDefinitions;
+                opt.Analyzer = analyzer;
+                opt.UseTaxonomyIndex = false;
+                opt.FacetsConfig = null;
+                opt.IndexValueTypesFactory = indexValueTypesFactory;
+                opt.Validator = validator;
+            };
+            return serviceCollection.AddExamineLuceneIndex<TIndex, FileSystemDirectoryFactory>(name, config);
+        }
+
+        /// <summary>
+        /// Registers an Examine index
+        /// </summary>
+        [Obsolete("To remove in Examine V5")]
+        public static IServiceCollection AddExamineLuceneIndex<TIndex, TDirectoryFactory>(
+            this IServiceCollection serviceCollection,
+            string name,
+            FieldDefinitionCollection fieldDefinitions = null,
+            Analyzer analyzer = null,
+            IValueSetValidator validator = null,
+            IReadOnlyDictionary<string, IFieldValueTypeFactory> indexValueTypesFactory = null)
+            where TIndex : LuceneIndex
+            where TDirectoryFactory : class, IDirectoryFactory
+        {
+            Action<ExamineLuceneIndexConfiguration<TIndex, TDirectoryFactory>> config = opt =>
+            {
+                opt.FieldDefinitions = fieldDefinitions;
+                opt.Analyzer = analyzer;
+                opt.UseTaxonomyIndex = false;
+                opt.FacetsConfig = null;
+                opt.IndexValueTypesFactory = indexValueTypesFactory;
+                opt.Validator = validator;
+            };
+            return serviceCollection.AddExamineLuceneIndex<TIndex, TDirectoryFactory>(name, config);
+        }
 
         /// <summary>
         /// Registers an Examine index
         /// </summary>
         public static IServiceCollection AddExamineLuceneIndex<TIndex, TDirectoryFactory>(
-            this IServiceCollection serviceCollection,
-            string name,
-            FieldDefinitionCollection? fieldDefinitions = null,
-            Analyzer? analyzer = null,
-            IValueSetValidator? validator = null,
-            IReadOnlyDictionary<string, IFieldValueTypeFactory>? indexValueTypesFactory = null,
-            FacetsConfig? facetsConfig = null)
-            where TIndex : LuceneIndex
-            where TDirectoryFactory : class, IDirectoryFactory
+           this IServiceCollection serviceCollection,
+           string name,
+           Action<ExamineLuceneIndexConfiguration<TIndex, TDirectoryFactory>> configuration)
+           where TIndex : LuceneIndex
+           where TDirectoryFactory : class, IDirectoryFactory
         {
+            var config = new ExamineLuceneIndexConfiguration<TIndex, TDirectoryFactory>(name);
+            configuration.Invoke(config);
+
             // This is the long way to add IOptions but gives us access to the
             // services collection which we need to get the dir factory
             serviceCollection.AddSingleton<IConfigureOptions<LuceneDirectoryIndexOptions>>(
@@ -68,12 +102,13 @@ namespace Examine
                     name,
                     (options) =>
                     {
-                        options.Analyzer = analyzer;
-                        options.Validator = validator;
-                        options.IndexValueTypesFactory = indexValueTypesFactory;
-                        options.FieldDefinitions = fieldDefinitions ?? options.FieldDefinitions;
+                        options.Analyzer = config.Analyzer;
+                        options.Validator = config.Validator;
+                        options.IndexValueTypesFactory = config.IndexValueTypesFactory;
+                        options.FieldDefinitions = config.FieldDefinitions ?? options.FieldDefinitions;
                         options.DirectoryFactory = services.GetRequiredService<TDirectoryFactory>();
-                        options.FacetsConfig = facetsConfig ?? new FacetsConfig();
+                        options.FacetsConfig = config.FacetsConfig ?? new FacetsConfig();
+                        options.UseTaxonomyIndex = config.UseTaxonomyIndex;
                     }));
 
             return serviceCollection.AddSingleton<IIndex>(services =>
@@ -90,47 +125,26 @@ namespace Examine
         }
 
         /// <summary>
-        /// Registers an Examine index
+        /// Registers a file system based Lucene Examine index
         /// </summary>
-        public static IServiceCollection AddExamineLuceneIndex<TIndex, TDirectoryFactory>(
+        public static IServiceCollection AddExamineLuceneIndex(
             this IServiceCollection serviceCollection,
             string name,
-            FieldDefinitionCollection? fieldDefinitions = null,
-            Analyzer? analyzer = null,
-            IValueSetValidator? validator = null,
-            IReadOnlyDictionary<string, IFieldValueTypeFactory>? indexValueTypesFactory = null,
-            FacetsConfig? facetsConfig = null,
-            bool useTaxonomyIndex = false)
-            where TIndex : LuceneIndex
-            where TDirectoryFactory : class, IDirectoryFactory
+            Action<ExamineLuceneIndexConfiguration<LuceneIndex, FileSystemDirectoryFactory>> configuration)
         {
-            // This is the long way to add IOptions but gives us access to the
-            // services collection which we need to get the dir factory
-            serviceCollection.AddSingleton<IConfigureOptions<LuceneDirectoryIndexOptions>>(
-                services => new ConfigureNamedOptions<LuceneDirectoryIndexOptions>(
-                    name,
-                    (options) =>
-                    {
-                        options.Analyzer = analyzer;
-                        options.Validator = validator;
-                        options.IndexValueTypesFactory = indexValueTypesFactory;
-                        options.FieldDefinitions = fieldDefinitions ?? options.FieldDefinitions;
-                        options.DirectoryFactory = services.GetRequiredService<TDirectoryFactory>();
-                        options.FacetsConfig = facetsConfig ?? new FacetsConfig();
-                        options.UseTaxonomyIndex = useTaxonomyIndex;
-                    }));
+            return serviceCollection.AddExamineLuceneIndex<LuceneIndex, FileSystemDirectoryFactory>(name, configuration);
+        }
 
-            return serviceCollection.AddSingleton<IIndex>(services =>
-            {
-                IOptionsMonitor<LuceneDirectoryIndexOptions> options
-                        = services.GetRequiredService<IOptionsMonitor<LuceneDirectoryIndexOptions>>();
-
-                TIndex index = ActivatorUtilities.CreateInstance<TIndex>(
-                    services,
-                    new object[] { name, options });
-
-                return index;
-            });
+        /// <summary>
+        /// Registers a file system based Lucene Examine index
+        /// </summary>
+        public static IServiceCollection AddExamineLuceneIndex<TIndex>(
+            this IServiceCollection serviceCollection,
+            string name,
+            Action<ExamineLuceneIndexConfiguration<TIndex, FileSystemDirectoryFactory>> configuration)
+            where TIndex : LuceneIndex
+        {
+            return serviceCollection.AddExamineLuceneIndex<TIndex, FileSystemDirectoryFactory>(name, configuration);
         }
 
         /// <summary>
@@ -163,34 +177,56 @@ namespace Examine
         /// <summary>
         /// Registers a lucene multi index searcher
         /// </summary>
+        [Obsolete("Will be removed in Examine V5")]
         public static IServiceCollection AddExamineLuceneMultiSearcher(
             this IServiceCollection serviceCollection,
             string name,
             string[] indexNames,
-            Analyzer analyzer = null,
-            FacetsConfig facetsConfig = null)
-            => serviceCollection.AddExamineSearcher<MultiIndexSearcher>(name, s =>
+            Analyzer analyzer = null)
+        {
+            var cfg = new Action<ExamineLuceneMultiSearcherConfiguration>(opt =>
             {
-                IEnumerable<IIndex> matchedIndexes = s.GetServices<IIndex>()
-                     .Where(x => indexNames.Contains(x.Name));
-
-                var parameters = new List<object>
-                {
-                    matchedIndexes,
-                };
-
-                if (facetsConfig != null)
-                {
-                    parameters.Add(facetsConfig);
-                }
-
-                if (analyzer != null)
-                {
-                    parameters.Add(analyzer);
-                }
-
-                return parameters;
+                opt.Analyzer = analyzer;
+                opt.FacetConfiguration = default;
             });
+            return AddExamineLuceneMultiSearcher(serviceCollection, name, indexNames, cfg);
+        }
+
+
+        /// <summary>
+        /// Registers a lucene multi index searcher
+        /// </summary>
+        public static IServiceCollection AddExamineLuceneMultiSearcher(
+            this IServiceCollection serviceCollection,
+            string name,
+            string[] indexNames,
+            Action<ExamineLuceneMultiSearcherConfiguration> configuration = null)
+        {
+            var cfg = new ExamineLuceneMultiSearcherConfiguration(name, indexNames);
+            configuration?.Invoke(cfg);
+            return serviceCollection.AddExamineSearcher<MultiIndexSearcher>(name, s =>
+                    {
+                        IEnumerable<IIndex> matchedIndexes = s.GetServices<IIndex>()
+                             .Where(x => indexNames.Contains(x.Name));
+
+                        var parameters = new List<object>
+                        {
+                    matchedIndexes,
+                        };
+
+                        if (cfg.FacetConfiguration != null)
+                        {
+                            parameters.Add(cfg.FacetConfiguration);
+                        }
+
+                        if (cfg.Analyzer != null)
+                        {
+                            parameters.Add(cfg.Analyzer);
+                        }
+
+                        return parameters;
+                    });
+        }
 
         /// <summary>
         /// Adds the Examine core services
