@@ -8,10 +8,17 @@ namespace Examine.Lucene.Directories
     public class GenericDirectoryFactory : DirectoryFactoryBase
     {
         private readonly Func<string, Directory> _factory;
-        private readonly Func<string, Directory> _taxonomyDirectoryFactory;
-        
+        private readonly Func<string, Directory>? _taxonomyDirectoryFactory;
+
         /// <inheritdoc/>
-        public GenericDirectoryFactory(Func<string, Directory> factory, Func<string, Directory> taxonomyDirectoryFactory = null)
+        [Obsolete("To remove in Examine V5")]
+        public GenericDirectoryFactory(Func<string, Directory> factory)
+        {
+            _factory = factory;
+        }
+
+        /// <inheritdoc/>
+        public GenericDirectoryFactory(Func<string, Directory> factory, Func<string, Directory> taxonomyDirectoryFactory)
         {
             _factory = factory;
             _taxonomyDirectoryFactory = taxonomyDirectoryFactory;
@@ -31,6 +38,11 @@ namespace Examine.Lucene.Directories
         /// <inheritdoc/>
         protected override Directory CreateTaxonomyDirectory(LuceneIndex luceneIndex, bool forceUnlock)
         {
+            if (_taxonomyDirectoryFactory is null)
+            {
+                throw new NullReferenceException("Taxonomy Directory factory is null. Use constructor with all parameters");
+            }
+
             Directory dir = _taxonomyDirectoryFactory(luceneIndex.Name + "taxonomy");
             if (forceUnlock)
             {
