@@ -43,11 +43,11 @@ namespace Examine.Lucene.Indexing
         /// <param name="taxonomyIndex"></param>
         /// Defaults to <see cref="CultureInvariantStandardAnalyzer"/>
         /// </param>
-        public FullTextType(string fieldName, ILoggerFactory logger, bool sortable = false, bool isFacetable = false, Analyzer? analyzer = null, bool taxonomyIndex = false)
+        public FullTextType(string fieldName, ILoggerFactory logger, bool isFacetable, bool taxonomyIndex, bool sortable, Analyzer analyzer)
             : base(fieldName, logger, true)
         {
             _sortable = sortable;
-            _analyzer = analyzer ?? new CultureInvariantStandardAnalyzer();
+            _analyzer = analyzer;
             _isFacetable = isFacetable;
             _taxonomyIndex = taxonomyIndex;
         }
@@ -61,7 +61,10 @@ namespace Examine.Lucene.Indexing
         /// Defaults to <see cref="CultureInvariantStandardAnalyzer"/>
         /// </param>
         /// <param name="sortable"></param>
+        [Obsolete("To be removed in Examine V5")]
+#pragma warning disable RS0027 // API with optional parameter(s) should have the most parameters amongst its public overloads
         public FullTextType(string fieldName, ILoggerFactory logger, Analyzer? analyzer = null, bool sortable = false)
+#pragma warning restore RS0027 // API with optional parameter(s) should have the most parameters amongst its public overloads
             : base(fieldName, logger, true)
         {
             _sortable = sortable;
