@@ -19,12 +19,33 @@ namespace Examine.Lucene.Providers
         /// </summary>
         /// <param name="name"></param>
         /// <param name="analyzer"></param>
+        [Obsolete("To remove in Examine V5")]
+        protected BaseLuceneSearcher(string name, Analyzer analyzer)
+            : base(name)
+        {
+            if (string.IsNullOrWhiteSpace(name))
+            {
+                throw new ArgumentException("Value cannot be null or whitespace.", nameof(name));
+            }
+
+            LuceneAnalyzer = analyzer;
+            _facetsConfig = GetDefaultFacetConfig();
+        }
+
+        /// <summary>
+        /// Constructor to allow for creating an indexer at runtime
+        /// </summary>
+        /// <param name="name"></param>
+        /// <param name="analyzer"></param>
         /// <param name="facetsConfig"></param>
         protected BaseLuceneSearcher(string name, Analyzer analyzer, FacetsConfig facetsConfig)
             : base(name)
         {
             if (string.IsNullOrWhiteSpace(name))
+            {
                 throw new ArgumentException("Value cannot be null or whitespace.", nameof(name));
+            }
+
             LuceneAnalyzer = analyzer;
             _facetsConfig = facetsConfig;
         }
@@ -40,8 +61,10 @@ namespace Examine.Lucene.Providers
         /// <returns></returns>
         public abstract ISearchContext GetSearchContext();
 
+#pragma warning disable RS0027 // API with optional parameter(s) should have the most parameters amongst its public overloads
         /// <inheritdoc />
-		public override IQuery CreateQuery(string? category = null, BooleanOperation defaultOperation = BooleanOperation.And)
+        public override IQuery CreateQuery(string? category = null, BooleanOperation defaultOperation = BooleanOperation.And)
+#pragma warning restore RS0027 // API with optional parameter(s) should have the most parameters amongst its public overloads
             => CreateQuery(category, defaultOperation, LuceneAnalyzer, new LuceneSearchOptions());
 
         /// <summary>
@@ -65,6 +88,16 @@ namespace Examine.Lucene.Providers
         {
             var sc = CreateQuery().ManagedQuery(searchText);
             return sc.Execute(options);
+        }
+
+        /// <summary>
+        /// Gets a FacetConfig with default configuration
+        /// </summary>
+        /// <returns>Facet Config</returns>
+        [Obsolete("To remove in Examine V5")]
+        public virtual FacetsConfig GetDefaultFacetConfig()
+        {
+            return new FacetsConfig();
         }
 
         /// <inheritdoc/>
