@@ -16,8 +16,27 @@ namespace Examine.Lucene.Search
         /// <param name="searchAfter">Optionally skip to results after the results from the previous search execution. Used for efficent deep paging.</param>
         /// <param name="trackDocumentMaxScore">Whether to track the maximum document score. For best performance, if not needed, leave false.</param>
         /// <param name="trackDocumentScores">Whether to Track Document Scores. For best performance, if not needed, leave false.</param>
+        [Obsolete("To remove in Examine 5.0")]
+#pragma warning disable RS0027 // API with optional parameter(s) should have the most parameters amongst its public overloads
+        public LuceneQueryOptions(int skip, int? take = null, SearchAfterOptions? searchAfter = null, bool trackDocumentScores = false, bool trackDocumentMaxScore = false)
+#pragma warning restore RS0027 // API with optional parameter(s) should have the most parameters amongst its public overloads
+            : base(skip, take)
+        {
+            TrackDocumentScores = trackDocumentScores;
+            TrackDocumentMaxScore = trackDocumentMaxScore;
+            SearchAfter = searchAfter;
+        }
+
+        /// <summary>
+        /// Constructor
+        /// </summary>
         /// <param name="facetSampling">Whether to apply Facet sampling to improve performance. If not required, leave null</param>
-        public LuceneQueryOptions(int skip, int? take = null, SearchAfterOptions? searchAfter = null, bool trackDocumentScores = false, bool trackDocumentMaxScore = false, LuceneFacetSamplingQueryOptions? facetSampling = null)
+        /// <param name="skip">Number of result documents to skip.</param>
+        /// <param name="take">Optional number of result documents to take.</param>
+        /// <param name="searchAfter">Optionally skip to results after the results from the previous search execution. Used for efficent deep paging.</param>
+        /// <param name="trackDocumentMaxScore">Whether to track the maximum document score. For best performance, if not needed, leave false.</param>
+        /// <param name="trackDocumentScores">Whether to Track Document Scores. For best performance, if not needed, leave false.</param>
+        public LuceneQueryOptions(LuceneFacetSamplingQueryOptions? facetSampling, int skip, int? take, SearchAfterOptions? searchAfter, bool trackDocumentScores, bool trackDocumentMaxScore)
             : base(skip, take)
         {
             TrackDocumentScores = trackDocumentScores;
