@@ -86,30 +86,30 @@ namespace Examine.Lucene
                 {FieldDefinitionTypes.FullTextSortable, name => new FullTextType(name, loggerFactory, defaultAnalyzer, true)},
                 {FieldDefinitionTypes.InvariantCultureIgnoreCase, name => new GenericAnalyzerFieldValueType(name, loggerFactory, new CultureInvariantWhitespaceAnalyzer())},
                 {FieldDefinitionTypes.EmailAddress, name => new GenericAnalyzerFieldValueType(name, loggerFactory, new EmailAddressAnalyzer())},
-                {FieldDefinitionTypes.FacetInteger, name => new Int32Type(name, loggerFactory, true, true, false)},
-                {FieldDefinitionTypes.FacetFloat, name => new SingleType(name, loggerFactory, true, true, false)},
-                {FieldDefinitionTypes.FacetDouble, name => new DoubleType(name, loggerFactory, true, true, false)},
-                {FieldDefinitionTypes.FacetLong, name => new Int64Type(name, loggerFactory, true, true, false)},
-                {FieldDefinitionTypes.FacetDateTime, name => new DateTimeType(name, loggerFactory, DateResolution.MILLISECOND, true, true, false)},
-                {FieldDefinitionTypes.FacetDateYear, name => new DateTimeType(name, loggerFactory, DateResolution.YEAR, true, true, false)},
-                {FieldDefinitionTypes.FacetDateMonth, name => new DateTimeType(name, loggerFactory, DateResolution.MONTH, true, true, false)},
-                {FieldDefinitionTypes.FacetDateDay, name => new DateTimeType(name, loggerFactory, DateResolution.DAY, true, true, false)},
-                {FieldDefinitionTypes.FacetDateHour, name => new DateTimeType(name, loggerFactory, DateResolution.HOUR, true, true, false)},
-                {FieldDefinitionTypes.FacetDateMinute, name => new DateTimeType(name, loggerFactory, DateResolution.MINUTE, true, true, false)},
-                {FieldDefinitionTypes.FacetFullText, name => new FullTextType(name, loggerFactory, true, true, defaultAnalyzer)},
-                {FieldDefinitionTypes.FacetFullTextSortable, name => new FullTextType(name, loggerFactory, true, true, defaultAnalyzer)},
-                {FieldDefinitionTypes.FacetTaxonomyInteger, name => new Int32Type(name, loggerFactory,  true,isFacetable: true, taxonomyIndex: true)},
-                {FieldDefinitionTypes.FacetTaxonomyFloat, name => new SingleType(name, loggerFactory,  true,isFacetable: true, taxonomyIndex: true)},
-                {FieldDefinitionTypes.FacetTaxonomyDouble, name => new DoubleType(name, loggerFactory,  true,isFacetable: true, taxonomyIndex: true)},
-                {FieldDefinitionTypes.FacetTaxonomyLong, name => new Int64Type(name, loggerFactory, true, isFacetable: true, taxonomyIndex: true)},
-                {FieldDefinitionTypes.FacetTaxonomyDateTime, name => new DateTimeType(name, loggerFactory, DateResolution.MILLISECOND, true, isFacetable: true, taxonomyIndex: true)},
-                {FieldDefinitionTypes.FacetTaxonomyDateYear, name =>  new DateTimeType(name, loggerFactory, DateResolution.YEAR, true,isFacetable: true, taxonomyIndex: true)},
-                {FieldDefinitionTypes.FacetTaxonomyDateMonth, name =>new DateTimeType(name, loggerFactory, DateResolution.MONTH, true,isFacetable: true, taxonomyIndex: true)},
-                {FieldDefinitionTypes.FacetTaxonomyDateDay, name =>new DateTimeType(name, loggerFactory, DateResolution.DAY, true,isFacetable: true, taxonomyIndex: true)},
-                {FieldDefinitionTypes.FacetTaxonomyDateHour, name => new DateTimeType(name, loggerFactory, DateResolution.HOUR, true,isFacetable: true, taxonomyIndex: true)},
-                {FieldDefinitionTypes.FacetTaxonomyDateMinute, name => new DateTimeType(name, loggerFactory, DateResolution.MINUTE, true, isFacetable: true, taxonomyIndex: true)},
-                {FieldDefinitionTypes.FacetTaxonomyFullText, name => new FullTextType(name, loggerFactory, false, true, defaultAnalyzer, taxonomyIndex: true)},
-                {FieldDefinitionTypes.FacetTaxonomyFullTextSortable, name => new FullTextType(name, loggerFactory, true, true, defaultAnalyzer,  taxonomyIndex: true)},
+                {FieldDefinitionTypes.FacetInteger, name => new Int32Type(name, true,false,loggerFactory, true)},
+                {FieldDefinitionTypes.FacetFloat, name => new SingleType(name, true, false, loggerFactory, true)},
+                {FieldDefinitionTypes.FacetDouble, name => new DoubleType(name,true, false, loggerFactory, true)},
+                {FieldDefinitionTypes.FacetLong, name => new Int64Type(name, true, false, loggerFactory, true)},
+                {FieldDefinitionTypes.FacetDateTime, name => new DateTimeType(name, true, true, false, loggerFactory, DateResolution.MILLISECOND)},
+                {FieldDefinitionTypes.FacetDateYear, name => new DateTimeType(name,true, true, false, loggerFactory, DateResolution.YEAR)},
+                {FieldDefinitionTypes.FacetDateMonth, name => new DateTimeType(name,true, true, false, loggerFactory, DateResolution.MONTH)},
+                {FieldDefinitionTypes.FacetDateDay, name => new DateTimeType(name, true, true, false, loggerFactory, DateResolution.DAY)},
+                {FieldDefinitionTypes.FacetDateHour, name => new DateTimeType(name,  true, true, false, loggerFactory, DateResolution.HOUR)},
+                {FieldDefinitionTypes.FacetDateMinute, name => new DateTimeType(name, true, true, false, loggerFactory, DateResolution.MINUTE)},
+                {FieldDefinitionTypes.FacetFullText, name => new FullTextType(name, loggerFactory, true, false, false, defaultAnalyzer ?? new CultureInvariantStandardAnalyzer())},
+                {FieldDefinitionTypes.FacetFullTextSortable, name => new FullTextType(name, loggerFactory, true, false,true, defaultAnalyzer ?? new CultureInvariantStandardAnalyzer())},
+                {FieldDefinitionTypes.FacetTaxonomyInteger, name => new Int32Type(name,true,true, loggerFactory, true)},
+                {FieldDefinitionTypes.FacetTaxonomyFloat, name => new SingleType(name,isFacetable: true, taxonomyIndex: true, loggerFactory,  true)},
+                {FieldDefinitionTypes.FacetTaxonomyDouble, name => new DoubleType(name, true, true, loggerFactory, true)},
+                {FieldDefinitionTypes.FacetTaxonomyLong, name => new Int64Type(name, isFacetable: true, taxonomyIndex: true, loggerFactory, true)},
+                {FieldDefinitionTypes.FacetTaxonomyDateTime, name => new DateTimeType(name,true, true, taxonomyIndex : true, loggerFactory, DateResolution.MILLISECOND)},
+                {FieldDefinitionTypes.FacetTaxonomyDateYear, name =>  new DateTimeType(name, true, true, taxonomyIndex : true, loggerFactory, DateResolution.YEAR)},
+                {FieldDefinitionTypes.FacetTaxonomyDateMonth, name => new DateTimeType(name, true, true, taxonomyIndex : true, loggerFactory, DateResolution.MONTH)},
+                {FieldDefinitionTypes.FacetTaxonomyDateDay, name => new DateTimeType(name, true, true, taxonomyIndex : true, loggerFactory, DateResolution.DAY)},
+                {FieldDefinitionTypes.FacetTaxonomyDateHour, name => new DateTimeType(name, true, isFacetable: true, taxonomyIndex: true, loggerFactory, DateResolution.HOUR)},
+                {FieldDefinitionTypes.FacetTaxonomyDateMinute, name => new DateTimeType(name, true, true, taxonomyIndex : true, loggerFactory, DateResolution.MINUTE)},
+                {FieldDefinitionTypes.FacetTaxonomyFullText, name => new FullTextType(name, loggerFactory, true, true, false, defaultAnalyzer ?? new CultureInvariantStandardAnalyzer())},
+                {FieldDefinitionTypes.FacetTaxonomyFullTextSortable, name => new FullTextType(name, loggerFactory, true, true, true, defaultAnalyzer ?? new CultureInvariantStandardAnalyzer())},
                 };
 
 
